Move form logic to subtraits
